# api/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/auth/me":{"get":{"operationId":"AuthController_decodingToken","summary":"","description":"get current login users informations","parameters":[],"responses":{"200":{"description":""}},"tags":["authentication"],"security":[{"bearer":[]}]}},"/api/auth/google":{"get":{"operationId":"AuthController_googleAuth","parameters":[],"responses":{"200":{"description":""}},"tags":["authentication"]}},"/api/auth/google/redirect":{"get":{"operationId":"AuthController_googleAuthRedirect","parameters":[],"responses":{"200":{"description":""}},"tags":["authentication"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","summary":"","description":"user login with credentials","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginPayload"}}}},"responses":{"201":{"description":""}},"tags":["authentication"]}},"/api/auth/logout":{"post":{"operationId":"AuthController_logout","summary":"","description":"user logout to revoke jwt","parameters":[],"responses":{"201":{"description":""}},"tags":["authentication"]}},"/api/auth/register":{"post":{"operationId":"AuthController_register","summary":"(ADMIN only)","description":"user register with input","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterPayload"}}}},"responses":{"201":{"description":""}},"tags":["authentication"]}},"/api/auth/refresh":{"post":{"operationId":"AuthController_refreshToken","summary":"","description":"refresh the access token","parameters":[],"responses":{"201":{"description":""}},"tags":["authentication"]}},"/api/auth/forgot":{"post":{"operationId":"AuthController_forgot","summary":"","description":"send reset password request","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPayload"}}}},"responses":{"201":{"description":""}},"tags":["authentication"]}},"/api/auth/reset":{"post":{"operationId":"AuthController_reset","summary":"","description":"reset user password","parameters":[{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPayload"}}}},"responses":{"201":{"description":""}},"tags":["authentication"]}},"/api/user":{"get":{"operationId":"UserController_search","summary":"(ADMIN only)","description":"get all users information","parameters":[{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]},"patch":{"operationId":"UserController_updateOwnUser","summary":"(USER only)","description":"update current user information","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPayload"}}}},"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/api/user/admin":{"get":{"operationId":"UserController_getAdmin","summary":"","description":"get all admin users information","parameters":[],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/api/user/{id}":{"get":{"operationId":"UserController_getUser","summary":"","description":"get a users information by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]},"patch":{"operationId":"UserController_updateUser","summary":"(ADMIN only)","description":"update a users information","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminUserPayload"}}}},"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserController_deleteUser","summary":"(ADMIN only)","description":"delete a users","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/api/checkin":{"get":{"operationId":"CheckinController_getCheckin","summary":"(USER only)","description":"get all checkin","parameters":[{"name":"fromDate","required":false,"in":"query","schema":{"format":"date-time","type":"string"}},{"name":"toDate","required":false,"in":"query","schema":{"format":"date-time","type":"string"}}],"responses":{"200":{"description":""}},"tags":["check in"],"security":[{"bearer":[]}]},"post":{"operationId":"CheckinController_checkin","summary":"(USER only)","description":"check in action","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckinoutPayload"}}}},"responses":{"201":{"description":""}},"tags":["check in"],"security":[{"bearer":[]}]}},"/api/checkin/today":{"get":{"operationId":"CheckinController_getCheckinToday","summary":"(USER only)","description":"get all checkin","parameters":[],"responses":{"200":{"description":""}},"tags":["check in"],"security":[{"bearer":[]}]}},"/api/checkout":{"post":{"operationId":"CheckoutController_checkout","summary":"(USER only)","description":"check out action","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckinoutPayload"}}}},"responses":{"201":{"description":""}},"tags":["check out"],"security":[{"bearer":[]}]}},"/api/comment":{"post":{"operationId":"CommentController_create","summary":"","description":"create comment","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentPayload"}}}},"responses":{"201":{"description":""}},"tags":["comment"],"security":[{"bearer":[]}]}},"/api/comment/{id}":{"get":{"operationId":"CommentController_getAllByTicketId","summary":"","description":"get all comments by ticket id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["comment"],"security":[{"bearer":[]}]},"patch":{"operationId":"CommentController_update","summary":"","description":"update a comment by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCommentPayload"}}}},"responses":{"200":{"description":""}},"tags":["comment"],"security":[{"bearer":[]}]},"delete":{"operationId":"CommentController_remove","summary":"","description":"delete a comment by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["comment"],"security":[{"bearer":[]}]}},"/api/ticket":{"get":{"operationId":"TicketController_getAll","summary":"(ADMIN only)","description":"get all tickets","parameters":[{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"default":"","type":"string"}},{"name":"ticketType","required":false,"in":"query","schema":{"default":"long term","type":"string"}},{"name":"ticketStatus","required":false,"in":"query","schema":{"default":"pending","type":"string"}},{"name":"sortField","required":false,"in":"query","schema":{"default":"createdAt","type":"string"}},{"name":"sortType","required":false,"in":"query","schema":{"default":true,"type":"boolean"}}],"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]},"post":{"operationId":"TicketController_createTicket","summary":"(USER only)","description":"create ticket","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTicketPayload"}}}},"responses":{"201":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]}},"/api/ticket/me":{"get":{"operationId":"TicketController_getAllByUserId","summary":"(USER only)","description":"get all tickets of current user","parameters":[{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"default":"","type":"string"}},{"name":"ticketType","required":false,"in":"query","schema":{"default":"long term","type":"string"}},{"name":"ticketStatus","required":false,"in":"query","schema":{"default":"pending","type":"string"}},{"name":"sortField","required":false,"in":"query","schema":{"default":"createdAt","type":"string"}},{"name":"sortType","required":false,"in":"query","schema":{"default":true,"type":"boolean"}}],"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]}},"/api/ticket/type":{"get":{"operationId":"TicketController_getTicketType","summary":"","description":"get all tickets type","parameters":[],"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]}},"/api/ticket/{id}":{"get":{"operationId":"TicketController_getByTicketId","summary":"","description":"get ticket details by ticket id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]},"patch":{"operationId":"TicketController_updateTicket","summary":"(USER only)","description":"update ticket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTicketPayload"}}}},"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]}},"/api/ticket/{id}/cancel":{"patch":{"operationId":"TicketController_cancelTicket","summary":"(USER only)","description":"cancel ticket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]}},"/api/ticket/{id}/reject":{"patch":{"operationId":"TicketController_rejectTicket","summary":"(ADMIN only)","description":"reject ticket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]}},"/api/ticket/{id}/approve":{"patch":{"operationId":"TicketController_approveTicket","summary":"(ADMIN only)","description":"approve ticket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["ticket"],"security":[{"bearer":[]}]}},"/api/notification":{"get":{"operationId":"NotificationController_getAllByRecipientId","summary":"","description":"get all notifications by user id","parameters":[],"responses":{"200":{"description":""}},"tags":["notification"],"security":[{"bearer":[]}]}},"/api/notification/{id}":{"patch":{"operationId":"NotificationController_updateReadNotification","summary":"","description":"mark is read a notification","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["notification"],"security":[{"bearer":[]}]}}},"info":{"title":"vCheck API","description":"API for timekeeping app","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:4005"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginPayload":{"type":"object","properties":{"email":{"type":"string","default":"test@vdtsol.com"},"password":{"type":"string","default":"Vcheck1@"}},"required":["email","password"]},"RegisterPayload":{"type":"object","properties":{"email":{"type":"string","default":"test@vdtsol.com"},"firstName":{"type":"string","default":"John"},"lastName":{"type":"string","default":"Wick"},"password":{"type":"string","default":"Vcheck1@"}},"required":["email","firstName","lastName","password"]},"ForgotPayload":{"type":"object","properties":{"email":{"type":"string","default":"test@vdtsol.com"}},"required":["email"]},"ResetPayload":{"type":"object","properties":{"password":{"type":"string"},"password2":{"type":"string"}},"required":["password","password2"]},"UserPayload":{"type":"object","properties":{"firstName":{"type":"string","default":"John"},"lastName":{"type":"string","default":"Wick"}},"required":["firstName","lastName"]},"AdminUserPayload":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"},"role":{"type":"object"}}},"CheckinoutPayload":{"type":"object","properties":{"image":{"type":"string"},"longitude":{"type":"string","default":"105.30304757837501"},"latitude":{"type":"string","default":"10.000207145497916"}},"required":["image","longitude","latitude"]},"CommentPayload":{"type":"object","properties":{"content":{"type":"string"},"ticketId":{"type":"number"}},"required":["content","ticketId"]},"UpdateCommentPayload":{"type":"object","properties":{"content":{"type":"string"}},"required":["content"]},"CreateTicketPayload":{"type":"object","properties":{"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"title":{"type":"string"},"content":{"type":"string"},"ticketType":{"type":"string","enum":["LONG_TERM","SHORT_TERM"]},"recipientId":{"type":"number"}},"required":["startDate","endDate","title","content","ticketType","recipientId"]},"UpdateTicketPayload":{"type":"object","properties":{"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"title":{"type":"string"},"content":{"type":"string"},"ticketType":{"type":"string","enum":["LONG_TERM","SHORT_TERM"]},"recipientId":{"type":"number"}}}}}}